requirements.txt
PyQt6>=6.4.0
PyQt6-Qt6==6.5.3
PyQt6-sip>=13.6,<14.0
Pillow>=9.0.0
send2trash==1.8.2
svglib==1.5.1
requests>=2.28.0 